pyproject.toml
[project]
name = "sunnytabbies-media-tools"
version = "0.1.0"
description = "Helper scripts to generate thumbnails and posters for SunnyTabbies gallery."
requires-python = ">=3.9"

dependencies = [
    "black>=25.9.0",
    "google-genai>=1.47.0",
    "pandas>=2.3.3",
    "pillow",
    "pydantic>=2.12.5",
    "pylance>=0.38.2",
    "pylint>=3.3.9",
    "python-dotenv>=1.2.1",
]
